fixed warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
libhif/hy-packageset.c:137:19: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
             if (index >= enabled ){
                       ^
libhif/hif-sack.c:585:37: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     if (key->name == 1 && key->size != data->repodataid)
                                     ^
libhif/hy-hth.c:216:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int i = 0; i < plist->len; ++i) {
                       ^
libhif/hy-hth.c:246:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int i = 0; i < plist->len; ++i) {
                       ^
libhif/hy-hth.c:267:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int i = 0; i < plist->len; ++i) {
                       ^
python/hawkey/iutil-py.c:49:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int i = 0; i < advisorylist->len; ++i) {
                       ^
python/hawkey/iutil-py.c:78:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int i = 0; i < advisorypkglist->len; ++i) {
                       ^
python/hawkey/iutil-py.c:106:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int i = 0; i < advisoryreflist->len; ++i) {
                       ^
python/hawkey/iutil-py.c:137:19: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (i = 0; i < plist->len; i++) {
                   ^
python/hawkey/iutil-py.c:195:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int i = 0; i < count; ++i) {
                       ^
python/hawkey/iutil-py.c:223:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int i = 0; i < count; ++i) {
                       ^
python/hawkey/query-py.c:277:27: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
         for (int i = 0; i < count; ++i) {
                           ^

+++ libhif/hif-sack.c
@@ -582,7 +582,7 @@ static int
 write_ext_updateinfo_filter(Repo *repo, Repokey *key, void *kfdata)
 {
     Repodata *data = kfdata;
-    if (key->name == 1 && key->size != data->repodataid)
+    if (key->name == 1 && (int) key->size != data->repodataid)
         return -1;
     return repo_write_stdkeyfilter(repo, key, 0);
 }
